iter1.16 | Изменил типы данных баланса на float64 и добавил логирование в пул
Типы данных для различных сущностей были изменены с int на float64, чтобы поддерживать более точные значения баланса и начислений. В пул обработки заказов добавлено множество логов для более детального отслеживания процесса.

Измененные файлы:
- internal/repositories/order.go
- internal/models/Account.go
- internal/payloads/order.go
- internal/repositories/accrual.go
- internal/payloads/accrual.go
- internal/ordercheck/pool.go
- internal/services/balanse.go
- internal/models/order.go
- internal/luna/luna.go
- internal/databse/migrations/migrations.go
- internal/helpers/helpers.go

+++ internal/models/Account.go
@@ -8,13 +8,13 @@ import (
 type Account struct {
 	ID          int64          `db:"id"`
 	UserID      int64          `db:"user_id"`
-	Difference  int            `db:"difference"`
+	Difference  float64        `db:"difference"`
 	OrderNumber sql.NullString `db:"order_number"`
 	CreatedAt   time.Time      `db:"created_at"`
 	UpdatedAt   time.Time      `db:"updated_at"`
 }
 
-func NewAccount(orderNumber sql.NullString, userID int64, difference int) *Account {
+func NewAccount(orderNumber sql.NullString, userID int64, difference float64) *Account {
 	return &Account{
 		Difference:  difference,
 		UserID:      userID,
@@ -25,6 +25,6 @@ func NewAccount(orderNumber sql.NullString, userID int64, difference int) *Accou
 }
 
 type Balance struct {
-	Current   int `db:"current" json:"current"`
-	Withdrawn int `db:"withdrawn" json:"withdrawn"`
+	Current   float64 `db:"current" json:"current"`
+	Withdrawn float64 `db:"withdrawn" json:"withdrawn"`
 }
